With large files the download size was to big and was causing the server to run out of memory. https://bugzilla.novell.com/show_bug.cgi?id=91091
git-svn-id: svn://svn.code.sf.net/p/ifolder/code/trunk@4827 db2431b9-df06-0410-971c-a38ea9a75485

+++ simias/src/Sync/ClientFile.cs
@@ -499,7 +499,7 @@ namespace Simias.Sync
 				return fileMap;
 			}
 			
-			sizeToSync = blockSize * serverHashMap.Length;
+			sizeToSync = (long)blockSize * (long)serverHashMap.Length;
 			long remainingBytes = node.Length % blockSize;
 			if (remainingBytes != 0)
 				sizeToSync = sizeToSync - blockSize + remainingBytes;
